refactor daily spend updates to use new Queue DS
